Refactor login page to improve code structure
